Combined ProcessHandler and Modeller into a single daemon.
Closes #70.
- Combined ProcessHandler and Modeller into a single daemon to avoid concurrency issues.
- Modified start/stop wades scripts.
