--- modeller_starter.py
@@ -1,5 +1,0 @@
-from src.main.modeller.Modeller import Modeller
-
-if __name__ == "__main__":
-    modeller = Modeller()
-    modeller.start()
