--- pshandler_starter.py
@@ -1,5 +1,0 @@
-from src.main.psHandler.ProcessHandler import ProcessHandler
-
-if __name__ == "__main__":
-    ps_handler = ProcessHandler()
-    ps_handler.start()
