+++ stop_wades.sh
@@ -1,2 +1,2 @@
 sudo kill "$(cat "/var/lib/wades/run/wades_reviver.pid")" || echo "Process is not running."
-sudo python3 ./stop_daemons.py
+sudo python3 ./wades stop
